cleanup, comment and reorganize code
